refactor: merge util files
- util 파일들 모음
