+++ Grenutil/MyEventEmitter/constants/errMsgs.js
