+++ Grenutil/MyEventEmitter/constants/type.js
