+++ Grenutil/MyPromise/constants/states.js
