+++ Grenutil/MyPromise/constants/types.js
